(Refacto) Renaming dataHandlers folders to data (following namespace name change).
