+++ gegelatilib/include/archive.h
@@ -6,7 +6,7 @@
 #include <deque>
 
 #include "mutator/rng.h"
-#include "dataHandlers/dataHandler.h"
+#include "data/dataHandler.h"
 #include "program/program.h"
 
 /**
