+++ gegelatilib/include/learn/learningEnvironment.h
@@ -3,7 +3,7 @@
 
 #include <cstdint>
 #include <vector>
-#include "dataHandlers/dataHandler.h"
+#include "data/dataHandler.h"
 
 namespace Learn {
