+++ test/mutatorTest.cpp
@@ -8,8 +8,8 @@
 #include "instructions/instruction.h"
 #include "instructions/addPrimitiveType.h"
 #include "instructions/multByConstParam.h"
-#include "dataHandlers/dataHandler.h"
-#include "dataHandlers/primitiveTypeArray.h"
+#include "data/dataHandler.h"
+#include "data/primitiveTypeArray.h"
 #include "program/program.h"
 #include "program/line.h"
 #include "program/programExecutionEngine.h"
