+++ test/tpgExecutionEngineTest.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 
-#include "dataHandlers/dataHandler.h"
-#include "dataHandlers/primitiveTypeArray.h"
+#include "data/dataHandler.h"
+#include "data/primitiveTypeArray.h"
 #include "instructions/addPrimitiveType.h"
 #include "instructions/multByConstParam.h"
 #include "program/program.h"
